refactor: DraftItem을 DraftList에서 별도의 파일로 분리
